refactor: relocate crm-api orphan adapter types to proper homes
Pure type relocation, zero behavior change.

- Move noopWorker/noopJobArgs, followUpSettingsRef, and
  deferredAggregatorReenqueuer wiring shims out of main.go into a new
  package-main sibling cmd/crm-api/adapters.go (siblings became legal in
  PR1's package-mode build sweep).
- Move CommsSourceContactLister out of the repository package (it carried
  no persistence logic, only source-pinning delegation) into adapters.go
  as the unexported commsSourceContactLister; nothing outside package main
  referenced it.
- Move meetingNoteLinkageTargetReader into package service next to the
  LinkageTargetReader interface it satisfies, exposed via a new
  NewLinkageTargetReader constructor returning the interface.
- Strip single-file-build-convention scaffolding comments; repoint the
  meeting_note.go doc comment and the test-adapter comment at the new homes.
- Prune now-unused main.go imports (stdsync, pgx).
